fix: error fallback UI
