--- src/app/(common)/gathering/error.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { useEffect } from "react";
-
-export default function Error({ error, reset }: { error: Error & { digest?: string }; reset: () => void }) {
-  useEffect(() => {
-    console.error(error);
-  }, [error]);
-
-  return (
-    <div>
-      <h2>Something went wrong!</h2>
-      <button onClick={() => reset()}>Try again</button>
-    </div>
-  );
-}
