+++ src/components/ErrorBoundary.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { Component, ErrorInfo, ReactNode } from "react";
+import Button from "@/components/Button";
 
 interface Props {
   children: ReactNode;
@@ -11,14 +12,11 @@ interface State {
   error: Error | null;
 }
 
-class ErrorBoundary extends Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      hasError: false,
-      error: null,
-    };
-  }
+export default class ErrorBoundary extends Component<Props, State> {
+  state: State = {
+    hasError: false,
+    error: null,
+  };
 
   static getDerivedStateFromError(error: Error): State {
     return { hasError: true, error };
@@ -34,9 +32,16 @@ class ErrorBoundary extends Component<Props, State> {
 
     if (hasError) {
       return (
-        <div>
-          <div>{error?.message ?? "알 수 없는 에러가 발생했습니다."}</div>
-          <button onClick={() => this.setState({ hasError: false })}>다시 시도하기</button>
+        <div className="flex h-[50vh] w-full items-center justify-center">
+          <div className="flex flex-col items-center gap-4">
+            <div className="flex flex-col items-center gap-1">
+              <div className="text-lg text-mono-11">죄송합니다.</div>
+              <div className="text-mono-9">{error?.message ?? "알 수 없는 에러가 발생했습니다."}</div>
+            </div>
+            <Button styleType={"outline"} onClick={() => this.setState({ hasError: false })} className="w-[120px]">
+              다시 시도하기
+            </Button>
+          </div>
         </div>
       );
     }
@@ -44,5 +49,3 @@ class ErrorBoundary extends Component<Props, State> {
     return children;
   }
 }
-
-export default ErrorBoundary;
